Docs/requirements.txt
asgiref==3.5.2
attrs==23.1.0
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.4
flake8==6.1.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
Pillow==9.4.0
pycodestyle==2.11.0
pydocstyle==6.3.0
pyflakes==3.1.0
pytz==2022.7.1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
snowballstemmer==2.2.0
sqlparse==0.4.2
uritemplate==4.1.1
